refactor: pack Strategy struct and use unchecked loop increments
Pack Strategy struct from 2 slots to 1 (address + uint96 cap).
Use SafeCast at write boundaries.

Convert all 13 bounded for-loop increments across BTCVault,
Loan, BTCVault__Validation, and StrategyStateLogic to
unchecked { ++i; } pattern.

+++ loan-provider/src/libraries/types/DataTypes.sol
@@ -610,14 +610,16 @@ library DataTypes {
      * @dev Contains strategy address and allocation cap for risk management
      */
     struct Strategy {
+        // ── Single slot (32B): strategy address + allocation cap
         /**
          * @notice Address of the tokenized strategy vault contract
          */
-        address strategy;
+        address strategy; // 20B
         /**
          * @notice Maximum number of assets that can be deposited in this strategy
+         * @dev Max ~79T in asset decimals. SafeCast enforced at write boundary.
          */
-        uint256 cap;
+        uint96 cap; // 12B
     }
 
     /**
